Context api 연동 완료
