+++ src/components/TodoList.jsx
@@ -1,11 +1,13 @@
+import { useTodos } from '../hooks/useTodos'
 import TodoItem from './TodoItem'
 
-function TodoList({ todos, toggleTodo, deleteTodo }) {
+function TodoList() {
+    const { todos } = useTodos()
     return (
         <>
             <ul className="mt-5">
                 {todos.map((todo) => (
-                    <TodoItem key={todo.id} todo={todo} toggleTodo={toggleTodo} deleteTodo={deleteTodo} />
+                    <TodoItem key={todo.id} todo={todo} />
                 ))}
             </ul>
         </>
