+++ src/main.jsx
@@ -1,10 +1,13 @@
 import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
 import App from './App'
+import { TodoProvider } from './context/TodoContext'
 import './index.css'
 
 createRoot(document.getElementById('root')).render(
     <StrictMode>
-        <App />
+        <TodoProvider>
+            <App />
+        </TodoProvider>
     </StrictMode>,
 )
